Likes repository created.
